重构项目结构 1、移除暂时没有明确存在价值的项目 2、后台由IIS托管的设计改为由WPF客户端程式直接承载 - 相关的项目变更 3、添加单元测试组（日后的所有类和方法应当有相关的单元测试）

+++ IWorld.Model/Relative.cs
@@ -38,7 +38,7 @@ namespace IWorld.Model
         /// </summary>
         /// <param name="nodeId"></param>
         /// <param name="nodeLayer"></param>
-        public Relative(long nodeId, long nodeLayer)
+        public Relative(int nodeId, int nodeLayer)
         {
             this.NodeId = nodeId;
             this.NodeLayer = nodeLayer;
